Extracted scene voice chats tracking service to maintain better SRP on StatusService
