+++ Explorer/Assets/DCL/VoiceChat/SceneVoiceChatTrackerService.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 59d86562f9a759c40b1f4a1eec4b14a2
