+++ Explorer/Assets/DCL/VoiceChat/VoiceChatContainer.cs
@@ -6,7 +6,6 @@ using DCL.VoiceChat.Services;
 using DCL.Web3.Identities;
 using DCL.WebRequests;
 using ECS;
-using ECS.SceneLifeCycle;
 using ECS.SceneLifeCycle.Realm;
 using System;
 
@@ -17,12 +16,11 @@ namespace DCL.VoiceChat
         private readonly IVoiceService rpcPrivateVoiceChatService;
         private readonly ICommunityVoiceService rpcCommunityVoiceChatService;
         private readonly PrivateVoiceChatCallStatusService privateVoiceChatCallStatusService;
-        private readonly CommunityVoiceChatCallStatusService communityVoiceChatCallStatusService;
         private readonly VoiceChatParticipantsStateService participantsStateService;
-        public readonly VoiceChatOrchestrator VoiceChatOrchestrator;
+        private readonly SceneVoiceChatTrackerService sceneVoiceChatTrackerService;
 
-        // Expose the community voice chat service for scene change system
-        public CommunityVoiceChatCallStatusService CommunityVoiceChatCallStatusService => communityVoiceChatCallStatusService;
+        public readonly CommunityVoiceChatCallStatusService CommunityVoiceChatCallStatusService;
+        public readonly VoiceChatOrchestrator VoiceChatOrchestrator;
 
         public VoiceChatContainer(
             IRPCSocialServices socialServiceRPC,
@@ -36,25 +34,25 @@ namespace DCL.VoiceChat
             IRealmData realmData)
         {
             rpcPrivateVoiceChatService = new RPCPrivateVoiceChatService(socialServiceRPC, socialServiceEventBus);
-            rpcCommunityVoiceChatService = new RPCCommunityVoiceChatService(socialServiceRPC, socialServiceEventBus, webRequestController);
-            participantsStateService = new VoiceChatParticipantsStateService(roomHub.VoiceChatRoom().Room(), identityCache);
-            communityVoiceChatCallStatusService = new CommunityVoiceChatCallStatusService(rpcCommunityVoiceChatService, notificationsBusController, parcelTrackerService, realmNavigator, realmData);
             privateVoiceChatCallStatusService = new PrivateVoiceChatCallStatusService(rpcPrivateVoiceChatService);
 
-            VoiceChatOrchestrator = new VoiceChatOrchestrator(
-                privateVoiceChatCallStatusService,
-                communityVoiceChatCallStatusService,
-                participantsStateService);
+            participantsStateService = new VoiceChatParticipantsStateService(roomHub.VoiceChatRoom().Room(), identityCache);
+
+            rpcCommunityVoiceChatService = new RPCCommunityVoiceChatService(socialServiceRPC, socialServiceEventBus, webRequestController);
+            sceneVoiceChatTrackerService = new SceneVoiceChatTrackerService(parcelTrackerService, realmNavigator, realmData);
+            CommunityVoiceChatCallStatusService = new CommunityVoiceChatCallStatusService(rpcCommunityVoiceChatService, notificationsBusController, sceneVoiceChatTrackerService);
+            VoiceChatOrchestrator = new VoiceChatOrchestrator(privateVoiceChatCallStatusService, CommunityVoiceChatCallStatusService, participantsStateService, sceneVoiceChatTrackerService);
         }
 
         public void Dispose()
         {
-            privateVoiceChatCallStatusService?.Dispose();
-            communityVoiceChatCallStatusService?.Dispose();
-            rpcPrivateVoiceChatService?.Dispose();
-            VoiceChatOrchestrator?.Dispose();
-            rpcCommunityVoiceChatService?.Dispose();
-            participantsStateService?.Dispose();
+            privateVoiceChatCallStatusService.Dispose();
+            CommunityVoiceChatCallStatusService.Dispose();
+            rpcPrivateVoiceChatService.Dispose();
+            VoiceChatOrchestrator.Dispose();
+            rpcCommunityVoiceChatService.Dispose();
+            participantsStateService.Dispose();
+            sceneVoiceChatTrackerService.Dispose();
         }
     }
 }
